Move pools to pages
